TTR-19: Added first version of the car wrapper and simple simulator.
The wrapper is not tested yet and does not handle car shutdown. Will be
added after testing with a car. Simulator is working, but does not give
out very good numbers. Again the simulator and car wrapper does not have
a good way of getting if the car is on or off. This will come later today.