TTR-24: Updated to support car module changes, fixed bugs and improved.
Added support for connected and Reconnected() in the car module. If not
connected, the core will now try a reconnect after every send.

Fixed time bug, instead of using time.localtime() which gives you a
struct, we now use datetime.now().

Instead of using a list and filling in data (caused key error if not all
data was not there) we now itterate over the data and fill the list.