TTR-37,TTR-48: Updated a lot, but should be working now!
Core was refactored and config was added to work with the core
Car was also updated to use config instead of hardcoding. Simulator needs
to be updated (or icar needs to be reverted).
dbwrapper was changed to only take one api address and returning status
code when sending. dbwrapper should be updated to handle if the sending
takes time.
gpspoller had minor bugfixes.

Core was also updated to send a trip start message.