TTR-34, TTR-31: Made dbwrapper more responsible
The DBWrapper needs to be responsible for data after the Core has issued
the send command. This is done by having the Core only focus about the
send command, and letting the DBWrapper store and buffer messages
automaticly. It also has to free the Core as soon as possible. This is
done by creating a thread for every send, which then has the
responsibility to send the message. If it fails multiple times, it will
buffer the messages, which will then be sent when another message is sent
successfully.